Merge separate webpack config and update package.json
Merges the separate production and development webpack
configs into one. To do this in a more understandable
and concise way, a few packages were installed, namely
webpack-config-utils. In addition to this, webpack config
validation was added as well as a lightweight progress
bar to reportstatus of builds.

Additionally, the package.json was updated to include
new build scripts for production and development. Previously,
webpack-dev-server ran using the node.js api, which has been
swapped out in favor of CLI usage.

Unnecessary packages were also stripped away to keep dependencies
down to the essentials.