Add long-term asset caching and NPM scripts
Generation of a manifest file was finalized to be inlined
in the head of the outputted html. This adds a content-dependent
cache-buster to each file. The entry point chunk containing the
bootstrapping code doesn't change its hash over time if its
dependencies are unchanged.

A copy plugin was also added to the Webpack config which moves
fonts into the build folder. Styles and scripts are moved by
Webpack after they have been processed.

The imagemin-cli and additional NPM build scripts were added to
facilitate image minification and relocation to the build directory.

Finally, the ESLint config was updated to use the new eslint-config-planetary
package for simplification of linting rule management.