Switch from SCSS to PostCSS
Replaces scss in favor of postcss. A few plugins have been
added to PostCSS to enable usage of next-gen CSS features.
The main workhorse here is cssnext, which allows color
manipulation, nesting, and a variety of other useful features
from the future css spec.

A few additional postcss plugins were added to provide additional
functionality. For the initial transition to postcss, the goal
was to keep additional plugins beyond cssnext to a minimum.

The plugins which were chosen for inclusion were postcss-mixins,
postcss-modular-scale, and postcss-import. While the names are
somewhat self-explanatory, there are a few quirks with modular-scale.

The modular-scale plugin does not seem to appropriately output
the correct sizes at this point, as the --ms-bases and --ms-ratios
does not have any effect. The outputted scale is modular, but the
ratio used appears to be 1.618 instead of the inputted 1.5.

Yet to be added is a stylelint config.