Fix modular scale and add stylelint config
A previous commit used a postcss plugin for achieving
modular scales which did not respond to changes in base
size or ratios. This plugin was replaced in favor
of the native modular scale css implementation which
behaves appropriately.

Additionally, a stylelint config was added as well as
a new build plugin for webpack to lint everything in
the assets/styles directory except for vendor stylesheets.